Possibility to add argument

+++ javaCompile.sh
@@ -3,12 +3,18 @@
 file=$1
 filename="${file%.*}"
 extension="${file##*.}"
-if [ $# -eq 1 ]; then # Test if the number of parameters is to 1
+if [ $# -eq 1 -o  $# -eq 2 ]; then # Test if the number of parameters is to 1
     if [ -f $file ]; then # Test if the file is really a file
         if [[ "$extension" == "java" ]]; then # And test if the extension is "java" and not "class"
-            rm *.class
-            javac $file
-            java $filename
+            if [ $# -eq 1 ];then
+                rm *.class
+                javac $file
+                java $filename
+            else 
+                rm *.class
+                javac $file
+                java $filename $2
+            fi
         else
             echo "Error don't forget to use the .java not .class"
         fi
